requirements.txt
hitherdither @ git+https://www.github.com/hbldh/hitherdither@0f3bbc44595aea2a20c4b96d8f536b05ac34e2a1
numpy==2.2.4
pillow==11.1.0
python-dotenv==1.0.1


# # hitherdither @ git+https://www.github.com/hbldh/hitherdither@0f3bbc44595aea2a20c4b96d8f536b05ac34e2a1
# # pip install git+https://www.github.com/hbldh/hitherdither 
# numpy==1.26.4
# pillow==10.3.0
# python-dotenv
